Use sbt-scalapb instead of our own protobuf plugin.
It is a protoc plugin, so does everything better. It also seems to be
more maintained, and they seem to have plans to support proto3.
