+++ project/idea-modules/generated.iml
@@ -9,6 +9,6 @@
     <orderEntry type="sourceFolder" forTests="false" />
     <orderEntry type="library" name="com.google.protobuf:protobuf-java:2.6.1" level="application" />
     <orderEntry type="library" name="scala-sdk-2.11.6" level="application" />
-    <orderEntry type="library" name="net.sandrogrzicic:scalabuff-runtime_2.11:1.4.0" level="application" />
+    <orderEntry type="library" name="com.trueaccord.scalapb:scalapb-runtime_2.11:0.4.12" level="application" />
   </component>
 </module>
