+++ project/plugins.sbt
@@ -14,8 +14,14 @@ addSbtPlugin("com.typesafe.sbt" % "sbt-scalariform" % "1.3.0")
 addSbtPlugin("org.scoverage" % "sbt-scoverage" % "1.0.1")
 addSbtPlugin("org.scoverage" % "sbt-coveralls" % "1.0.0.BETA1")
 
+// Scala protobuf support.
+addSbtPlugin("com.trueaccord.scalapb" % "sbt-scalapb" % "0.4.12")
+
 // Scala compiler options for SBT code.
 scalacOptions ++= Seq("-feature", "-deprecation")
 
 // Dependencies for SBT code.
-libraryDependencies += "commons-io" % "commons-io" % "2.4"
+libraryDependencies ++= Seq(
+  "commons-io" % "commons-io" % "2.4",
+  "com.google.protobuf" % "protobuf-java" % "2.6.1"
+)
