+++ project/src/main/scala/sbt/tox4j/CodeFormat.scala
@@ -1,4 +1,4 @@
-package src.main.scala
+package sbt.tox4j
 
 import com.typesafe.sbt.SbtScalariform.{ ScalariformKeys, scalariformSettings }
 import sbt._
