+++ project/src/main/scala/sbt/tox4j/Tox4jLibraryBuild.scala
@@ -1,4 +1,4 @@
-package src.main.scala
+package sbt.tox4j
 
 import sbt.Keys._
 import sbt._
